Some refactoring and cleanup ! :D
