--- src/main/java/com/skitch/trason/addon/elements/events/bukkit/BridgeEventOffLive.java
@@ -1,31 +1,0 @@
-package com.skitch.trason.addon.elements.events.bukkit;
-
-import com.github.twitch4j.events.ChannelGoOfflineEvent;
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
-import org.jetbrains.annotations.NotNull;
-
-public class BridgeEventOffLive extends Event {
-
-    private static final HandlerList handlers = new HandlerList();
-
-    private final ChannelGoOfflineEvent event;
-    private final String ChannelName;
-
-    public BridgeEventOffLive(ChannelGoOfflineEvent event) {
-        this.event = event;
-        this.ChannelName = event.getChannel().getName();
-    }
-
-    public ChannelGoOfflineEvent getEvent() {
-        return event;
-    }
-
-    @NotNull
-    @Override
-    public HandlerList getHandlers() {
-        return handlers;
-    }
-
-    public static HandlerList getHandlerList() { return handlers; }
-}
