+++ src/main/java/com/trason/skitch/elements/events/bukkit/BridgeEventGiftSub.java
@@ -1,0 +1,30 @@
+package com.trason.skitch.elements.events.bukkit;
+
+import com.github.twitch4j.chat.events.channel.GiftSubscriptionsEvent;
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.jetbrains.annotations.NotNull;
+
+public class BridgeEventGiftSub extends Event {
+
+    private static final HandlerList handlers = new HandlerList();
+
+    private final GiftSubscriptionsEvent event;
+
+    public BridgeEventGiftSub(GiftSubscriptionsEvent event) {
+        this.event = event;
+    }
+
+    public GiftSubscriptionsEvent getEvent() {
+        return event;
+    }
+
+    @NotNull
+    @Override
+    public HandlerList getHandlers() {
+        return handlers;
+    }
+
+    public static HandlerList getHandlerList() { return handlers; }
+
+}
