+++ src/main/java/com/trason/skitch/elements/expressions/ExprStreamTitle.java
@@ -1,0 +1,63 @@
+package com.trason.skitch.elements.expressions;
+
+import ch.njol.skript.Skript;
+import ch.njol.skript.lang.Expression;
+import ch.njol.skript.lang.ExpressionType;
+import ch.njol.skript.lang.SkriptParser.ParseResult;
+import ch.njol.skript.lang.util.SimpleExpression;
+import ch.njol.util.Kleenean;
+import com.github.twitch4j.helix.domain.Stream;
+import com.github.twitch4j.helix.domain.StreamList;
+import com.trason.skitch.elements.events.bukkit.*;
+import org.bukkit.event.Event;
+
+import java.util.Collections;
+
+import static com.trason.skitch.elements.effects.EffLoginTwitchBot.client;
+
+public class ExprStreamTitle extends SimpleExpression<String> {
+
+    static {
+        Skript.registerExpression(ExprStreamTitle.class, String.class, ExpressionType.SIMPLE,
+            "[event-]streamtitle");
+    }
+
+    @Override
+    public boolean init(Expression<?>[] exprs, int matchedPattern, Kleenean isDelayed, ParseResult parseResult) {
+        return true;
+    }
+
+    @Override
+    protected String[] get(Event event) {
+        if (event instanceof BridgeEventGoLive) {
+            String title = ((BridgeEventGoLive) event).getEvent().getStream().getTitle();
+            return new String[]{title};
+        }
+        else if (event instanceof BridgeEventChat) {
+            StreamList list = client.getHelix().getStreams(null, null, null, 1,
+                null, null, null,
+                Collections.singletonList(((BridgeEventChat) event).getEvent().getChannel().getName())).execute();
+            Stream str = list.getStreams().get(0);
+            String title = str.getTitle();
+            return new String[]{title};
+        }
+        else
+            return new String[0];
+    }
+
+    @Override
+    public boolean isSingle() {
+        return true;
+    }
+
+    @Override
+    public Class<? extends String> getReturnType() {
+        return String.class;
+    }
+
+    @Override
+    public String toString(Event e, boolean debug) {
+        return "event-streamtitle";
+    }
+
+}
